# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Equipment Tracker API",
    "description": "API for managing inventory and customer write operations.",
    "version": "1.0.0"
  },
  "basePath": "/",
  "paths": {
    "/inventory": {
      "get": {
        "summary": "List all inventory items",
        "description": "Returns every inventory item in the collection.",
        "responses": {
          "200": {
            "description": "Inventory items returned"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Inventory"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Inventory item created"
          },
          "400": {
            "description": "Invalid inventory data"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/customers": {
      "get": {
        "summary": "List all customers",
        "description": "Returns every customer in the collection.",
        "responses": {
          "200": {
            "description": "Customers returned"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Customer"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Customer created"
          },
          "400": {
            "description": "Invalid customer data"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "definitions": {
    "Inventory": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Excavator"
        },
        "description": {
          "type": "string",
          "example": "Tracked excavator for construction work."
        },
        "category": {
          "type": "string",
          "example": "Heavy Machinery"
        },
        "serialNumber": {
          "type": "string",
          "example": "EXC-001"
        },
        "dailyRate": {
          "type": "number",
          "example": 1500
        },
        "status": {
          "type": "string",
          "example": "available"
        },
        "condition": {
          "type": "string",
          "example": "Good"
        }
      }
    },
    "Customer": {
      "type": "object",
      "properties": {
        "firstName": {
          "type": "string",
          "example": "John"
        },
        "lastName": {
          "type": "string",
          "example": "Doe"
        },
        "email": {
          "type": "string",
          "example": "john@example.com"
        },
        "phone": {
          "type": "string",
          "example": "555-123-4567"
        }
      }
    }
  }
}